Finished assignment.
replaced tga image with stb image write for bmp files.
removed unused classes.
Removed camera to make code more direct
added mymodel.h to match assignment criteria.
Added parametric sphere intersection method
made polygon pnpoly code more readable

+++ src/Vector3.cpp
@@ -89,7 +89,7 @@ float Vector3::distanceSquared(const Vector3& v) const
 	return (*this - v).getSquaredLength();
 }
 
-float Vector3::getLargestComponentIndex() const
+int Vector3::getLargestComponentIndex() const
 {
 	float a_x = abs(x);
 	float a_y = abs(y);
